WIP : shadowmap en forward shading, ne compile pas

+++ S5engine/src/core/resources/glslshader.cpp
@@ -124,7 +124,7 @@ void GLSLShaderProgramFactory::load(ResourceData *resource)
 
 	QDomNodeList nodes = programs.at(0).childNodes();
 
-	for(int i=0 ; i < nodes.length() ; i++)
+	for(int i=0 ; i < (int)nodes.length() ; i++)
 	{
 		QString tag = nodes.at(i).nodeName();
